new feature: Compute 'tensor product' of multicurves
The 'tensor product' of multicurves corresponds to tangle sums (stacking
two Conway tangles on top of each other to obtain a new Conway tangle).
It allows us to compute the invariants of tangle sums from already
computed multicurves. There is also an option to only take non-compact
components as input, which is useful for studying concordance.